greens function tests pass
