+++ src/MaterialsScienceTools.jl
@@ -4,7 +4,7 @@ include("common.jl")
 
 include("CLE.jl")
 
-# include("GreensFunctions.jl")
+include("GreensFunctions.jl")
 
 
 # include("bbs.jl")
